QuickFit3/imFCS:   - error corrections in correlators   - added distance CCF
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@1129 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/imagingfcs/qfrdrimagingfcs_data.cpp
@@ -131,9 +131,9 @@ bool QFRDRImagingFCSData::loadOverview(QString filename) {
 
             qDebug()<<"TIFF: "<<nx<<"x"<<ny<<"    data: "<<width<<"x"<<height;
 
-            if (width*height<=nx*ny) {
-                for (uint32_t y=0; y<height; y++) {
-                    for (uint32_t x=0; x<width; x++) {
+            if (width*height<=(int64_t)nx*(int64_t)ny) {
+                for (int32_t y=0; y<height; y++) {
+                    for (int32_t x=0; x<width; x++) {
                         overview[y*width+x]=ovw[y*nx+x];
                     }
